+++ client/src/effect/laser_beam.rs
@@ -1,5 +1,5 @@
 // STD Dependencies -----------------------------------------------------------
-use std::f64::consts;
+use std::f32::consts;
 
 
 // External Dependencies ------------------------------------------------------
@@ -15,7 +15,7 @@ use shared::color::{Color, ColorName};
 
 // Laser Effect ---------------------------------------------------------------
 pub struct LaserBeam {
-    pub line: [f64; 4],
+    pub line: [f32; 4],
     color_dark: [f32; 4],
     color_light: [f32; 4],
     start: u64,
@@ -24,7 +24,7 @@ pub struct LaserBeam {
 
 impl LaserBeam {
 
-    pub fn new(color: ColorName, line: [f64; 4], duration: u64) -> LaserBeam {
+    pub fn new(color: ColorName, line: [f32; 4], duration: u64) -> LaserBeam {
         LaserBeam {
             line: line,
             color_dark: Color::from_name(color).darken(0.5).into_f32(),
@@ -34,7 +34,7 @@ impl LaserBeam {
         }
     }
 
-    pub fn from_point(color: ColorName, x: f64, y: f64, r: f64, d: f64, l: f64, duration: u64) -> LaserBeam {
+    pub fn from_point(color: ColorName, x: f32, y: f32, r: f32, d: f32, l: f32, duration: u64) -> LaserBeam {
         LaserBeam::new(color, [
             x + r.cos() * d, y + r.sin() * d,
             x + r.cos() * (d + l), y + r.sin() * (d + l)
@@ -54,8 +54,8 @@ impl Effect for LaserBeam {
 
         let context = camera.context();
         let exp = renderer.t() - self.start;
-        let u = ((1.0 / self.duration as f64) * exp as f64).min(1.0).max(0.0);
-        let a = (0.35 + u * 0.5) as f32;
+        let u = ((1.0 / self.duration as f32) * exp as f32).min(1.0).max(0.0);
+        let a = 0.35 + u * 0.5;
 
         renderer.set_color([
             self.color_dark[0],
